--- controllers/placeholder.js
@@ -1,1 +1,0 @@
-//REMOVE THIS FILE ONCE WHEN THE ACTUAL CONTROLLER FILE IS ADDED.
